requirements.txt
# PingAn Life Returns Analyzer - Dependencies
# 平安寿险收益分析器 - 依赖包

# Data Processing
pandas>=1.5.0
numpy>=1.23.0
pyarrow>=10.0.0

# Financial Calculations
numpy-financial>=1.0.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.10.0

# Report Generation
openpyxl>=3.0.10
xlsxwriter>=3.0.3

# Configuration
pyyaml>=6.0

# Utilities
python-dateutil>=2.8.0
scipy>=1.9.0
